# packages/api-client/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "CreateExpertRequestDto": {
        "additionalProperties": false,
        "properties": {
          "affiliation": {
            "anyOf": [
              {
                "maxLength": 240,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "bio": {
            "anyOf": [
              {
                "maxLength": 4000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "credentials": {
            "anyOf": [
              {
                "maxLength": 500,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "familyName": {
            "maxLength": 80,
            "minLength": 1,
            "type": "string"
          },
          "givenName": {
            "maxLength": 80,
            "minLength": 1,
            "type": "string"
          },
          "patronymic": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "professionalRole": {
            "anyOf": [
              {
                "maxLength": 160,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "userId": {
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "type": "string"
          }
        },
        "required": ["familyName", "givenName"],
        "type": "object"
      },
      "EligibleExpertUserListDto": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "items": {
              "additionalProperties": false,
              "properties": {
                "displayName": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "id": {
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                  "type": "string"
                },
                "identifier": {
                  "minLength": 1,
                  "type": "string"
                }
              },
              "required": ["id", "displayName", "identifier"],
              "type": "object"
            },
            "type": "array"
          },
          "page": {
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991,
            "type": "integer"
          },
          "pageSize": {
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991,
            "type": "integer"
          },
          "total": {
            "maximum": 9007199254740991,
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": ["data", "total", "page", "pageSize"],
        "type": "object"
      },
...
      "ExpertAdminDetailDto": {
        "properties": {
          "affiliation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "bio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "contentRemovedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "createdAt": {
            "type": "string"
          },
          "credentials": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "familyName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "firstPublishedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "givenName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "initials": {
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "patronymic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "photoUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "professionalRole": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "slug": {
            "maxLength": 80,
            "minLength": 1,
            "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$",
            "type": "string"
          },
          "status": {
            "enum": ["draft", "published", "retired"],
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "userId": {
            "anyOf": [
              {
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "version": {
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991,
            "type": "integer"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "familyName",
          "givenName",
          "patronymic",
          "userId",
          "professionalRole",
          "credentials",
          "affiliation",
          "bio",
          "photoUrl",
          "initials",
          "status",
          "firstPublishedAt",
          "contentRemovedAt",
          "version",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
      "ExpertAdminListDto": {
        "properties": {
          "data": {
            "items": {
              "properties": {
                "id": {
                  "type": "string"
                },
                "name": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "professionalRole": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "slug": {
                  "maxLength": 80,
                  "minLength": 1,
                  "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$",
                  "type": "string"
                },
                "status": {
                  "enum": ["draft", "published", "retired"],
                  "type": "string"
                },
                "updatedAt": {
                  "type": "string"
                },
                "version": {
                  "exclusiveMinimum": 0,
                  "maximum": 9007199254740991,
                  "type": "integer"
                }
              },
              "required": [
                "id",
                "slug",
                "name",
                "professionalRole",
                "status",
                "version",
                "updatedAt"
              ],
              "type": "object"
            },
            "type": "array"
          },
          "page": {
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991,
            "type": "integer"
          },
          "pageSize": {
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991,
            "type": "integer"
          },
          "total": {
            "maximum": 9007199254740991,
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": ["data", "total", "page", "pageSize"],
        "type": "object"
      },
...
      "UpdateExpertRequestDto": {
        "additionalProperties": false,
        "properties": {
          "affiliation": {
            "anyOf": [
              {
                "maxLength": 240,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "bio": {
            "anyOf": [
              {
                "maxLength": 4000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "credentials": {
            "anyOf": [
              {
                "maxLength": 500,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "familyName": {
            "maxLength": 80,
            "minLength": 1,
            "type": "string"
          },
          "givenName": {
            "maxLength": 80,
            "minLength": 1,
            "type": "string"
          },
          "mediaAction": {
            "const": "clear",
            "type": "string"
          },
          "patronymic": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "professionalRole": {
            "anyOf": [
              {
                "maxLength": 160,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "userId": {
            "anyOf": [
              {
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "",
    "title": "DS Platform API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/experts": {
      "get": {
        "operationId": "ExpertsAdminController_list",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpertAdminListDto"
                }
              }
            },
            "description": ""
          }
        },
        "tags": ["ExpertsAdmin"]
      },
      "post": {
        "operationId": "ExpertsAdminController_create",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExpertRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpertAdminDetailDto"
                }
              }
            },
            "description": ""
          }
        },
        "tags": ["ExpertsAdmin"]
      }
    },
    "/v1/admin/experts/{id}": {
      "get": {
        "operationId": "ExpertsAdminController_detail",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpertAdminDetailDto"
                }
              }
            },
            "description": ""
          }
        },
        "tags": ["ExpertsAdmin"]
      },
      "patch": {
        "operationId": "ExpertsAdminController_update",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExpertRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpertAdminDetailDto"
                }
              }
            },
            "description": ""
          }
        },
        "tags": ["ExpertsAdmin"]
      }
    },
    "/v1/admin/experts/{id}/publish": {
      "post": {
        "operationId": "ExpertsAdminController_publish",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["ExpertsAdmin"]
      }
    },
    "/v1/admin/experts/eligible-users": {
      "get": {
        "operationId": "ExpertsAdminController_eligibleUsers",
        "parameters": [
          {
            "in": "query",
            "name": "currentExpertId",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "maxLength": 254,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EligibleExpertUserListDto"
                }
              }
            },
            "description": ""
          }
        },
        "tags": ["ExpertsAdmin"]
      }
    },
...
  },
  "servers": [],
...
}